remove unnecessary transformer utils
