+++ mmpose/models/utils/__init__.py
@@ -1,6 +1,9 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .ckpt_convert import pvt_convert
 from .smpl import SMPL
-from .transformer import PatchEmbed, nchw_to_nlc, nlc_to_nchw
+from .transformer import PatchEmbed, PatchMerging, nchw_to_nlc, nlc_to_nchw
 
-__all__ = ['SMPL', 'PatchEmbed', 'nchw_to_nlc', 'nlc_to_nchw', 'pvt_convert']
+__all__ = [
+    'SMPL', 'PatchEmbed', 'nchw_to_nlc', 'nlc_to_nchw', 'pvt_convert',
+    'PatchMerging'
+]
